Add files via upload
Objective: Banking Information system is a windows based application. This project mainly deals with managing accounts and their related operations. Task involve in this project are opening the user account, recording the account holders, transactions, depositing and withdrawing amounts. The project also includes the use of inno-setup compiler to package the entire project into a single executable for easy distribution and installing.
Tools Used: Python and its modules, Tkinter, Mysql, Sqlite (DB Browser)
Outcome: The project titled as “Banking Automation” is a windows based application. This software developed with scalability in mind. Additional modules can be easily added when necessary. The software is developed with modular approach to deal with managing accounts and their related operations. Task involved in this project are opening the user accouts, recording the account holder’s transactions, depositing and withdrawing amount. All modules are in the system has been tested with valid data and invalid data.